Research.py
from duckduckgo_search import DDGS
import requests
from bs4 import BeautifulSoup

def research_water_hardness(sector):
    """
    1. Searches the web for the water quality report for the sector.
    2. Extracts hardness data (this is a simplified logic for demonstration).
    """
    print(f"Agent: Researching {sector}...")
    
    # 1. Search for the official report
    query = f"water quality hardness report for {sector} sector UK"
    with DDGS() as ddgs:
        # Get the first official-looking result
        results = list(ddgs.text(query, max_results=1))
        
    if not results:
        return {"ppm_value": 0, "classification": "Unknown"}

    # 2. Logic to parse (In a real scenario, you would fetch the URL and parse)
    # For now, we simulate the extraction of a numeric value
    # You would typically add: response = requests.get(results[0]['href'])
    
    return {"ppm_value": 200, "classification": "Hard"}
