requirements.txt
mysql-connector-python
python-dotenv
duckduckgo-search
beautifulsoup4
requests
streamlit
